Move check_osd_log_exist_on_rook_ceph_operator_pod and get_last_log_time_date functions to helpers file
